memory/pyproject.toml
[project]
name = "clingysocks-memory"
version = "0.1.0"
description = "Relational memory harvest system with two-pass extraction pipeline"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109",
    "uvicorn>=0.27",
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9",
    "pgvector>=0.2",
    "litellm>=1.0",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27",
    "cryptography>=42.0",
    "bcrypt>=4.0",
    "PyJWT>=2.8",
    "mcp[cli]>=0.1",
    "rich>=13.0",
    "google-genai>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.setuptools.packages.find]
where = ["src"]
